[Data] Clean up duplicate torch utils in air (#60605)
## Description
- Remove duplicate function in
`‎python/ray/air/_internal/torch_utils.py‎`
- Update the import path to ray data
- Move test for `TestConvertPandasToTorch` to ray data

## Related issues
> Link related issues: "Fixes #1234", "Closes #1234", or "Related to
#1234".

## Additional information
> Optional: Add implementation details, API changes, usage examples,
screenshots, etc.

---------
