+++ python/ray/train/torch/torch_predictor.py
@@ -4,7 +4,7 @@ from typing import TYPE_CHECKING, Dict, Optional, Union
 import numpy as np
 import torch
 
-from ray.air._internal.torch_utils import convert_ndarray_batch_to_torch_tensor_batch
+from ray.data.util.torch_utils import convert_ndarray_batch_to_torch_tensor_batch
 from ray.train._internal.dl_predictor import DLPredictor
 from ray.train.predictor import DataBatchType
 from ray.train.torch import TorchCheckpoint
